requirements.txt
asgiref==3.2.10
backcall==0.2.0
cffi==1.14.3
cryptography==3.2.1
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.2
django-environ==0.4.5
django-extensions==3.0.9
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.1
gunicorn==20.0.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.2
python-dotenv==0.15.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
traitlets==4.3.3
wcwidth==0.2.5
whitenoise==5.2.0
